+++ sdk/core/Azure.Core/src/Pipeline/HttpPipeline.cs
@@ -54,7 +54,8 @@ namespace Azure.Core.Pipeline
             policies ??= Array.Empty<HttpPipelinePolicy>();
 
             var all = new HttpPipelinePolicy[policies.Length + 1];
-            all[policies.Length] = new HttpPipelineTransportPolicy(_transport);
+            all[policies.Length] = new HttpPipelineTransportPolicy(_transport,
+                ClientDiagnostics.CreateMessageSanitizer(new DiagnosticsOptions()));
             policies.CopyTo(all, 0);
 
             _pipeline = all;