+++ sdk/core/Azure.Core/tests/HttpPipelineTests.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using Azure.Core.Pipeline;
 using Azure.Core.TestFramework;
@@ -12,6 +13,25 @@ namespace Azure.Core.Tests
 {
     public class HttpPipelineTests
     {
+        [Test]
+        public async Task CanBuildPipelineAndSendMessage()
+        {
+            var mockTransport = new MockTransport(
+                new MockResponse(500),
+                new MockResponse(1));
+
+            var pipeline = new HttpPipeline(mockTransport, new[] {
+                new RetryPolicy(RetryMode.Exponential, TimeSpan.Zero, TimeSpan.Zero, 5)
+            }, responseClassifier: new CustomResponseClassifier());
+
+            Request request = pipeline.CreateRequest();
+            request.Method = RequestMethod.Get;
+            request.Uri.Reset(new Uri("https://contoso.a.io"));
+            Response response = await pipeline.SendRequestAsync(request, CancellationToken.None);
+
+            Assert.AreEqual(1, response.Status);
+        }
+
         [Test]
         public async Task DoesntDisposeRequestInSendRequestAsync()
         {
@@ -26,10 +46,6 @@ namespace Azure.Core.Tests
 
             Assert.False(request.IsDisposed);
             Assert.False(response.IsDisposed);
-        }
-
-        private class TestOptions : ClientOptions
-        {
         }
 
         [Test]
@@ -222,6 +238,54 @@ namespace Azure.Core.Tests
             Assert.IsTrue(throws);
         }
 
+        [Test]
+        public async Task PipelineSetsResponseIsErrorTrue()
+        {
+            var mockTransport = new MockTransport(
+                new MockResponse(500));
+
+            var pipeline = new HttpPipeline(mockTransport);
+
+            Request request = pipeline.CreateRequest();
+            request.Method = RequestMethod.Get;
+            request.Uri.Reset(new Uri("https://contoso.a.io"));
+            Response response = await pipeline.SendRequestAsync(request, CancellationToken.None);
+
+            Assert.IsTrue(response.IsError);
+        }
+
+        [Test]
+        public async Task PipelineSetsResponseIsErrorFalse()
+        {
+            var mockTransport = new MockTransport(
+                new MockResponse(200));
+
+            var pipeline = new HttpPipeline(mockTransport);
+
+            Request request = pipeline.CreateRequest();
+            request.Method = RequestMethod.Get;
+            request.Uri.Reset(new Uri("https://contoso.a.io"));
+            Response response = await pipeline.SendRequestAsync(request, CancellationToken.None);
+
+            Assert.IsFalse(response.IsError);
+        }
+
+        [Test]
+        public async Task CustomClassifierSetsResponseIsError()
+        {
+            var mockTransport = new MockTransport(
+                new MockResponse(404));
+
+            var pipeline = new HttpPipeline(mockTransport, responseClassifier: new CustomResponseClassifier());
+
+            Request request = pipeline.CreateRequest();
+            request.Method = RequestMethod.Get;
+            request.Uri.Reset(new Uri("https://contoso.a.io"));
+            Response response = await pipeline.SendRequestAsync(request, CancellationToken.None);
+
+            Assert.IsFalse(response.IsError);
+        }
+
         #region Helpers
         public class AddHeaderPolicy : HttpPipelineSynchronousPolicy
         {
@@ -237,6 +301,28 @@ namespace Azure.Core.Tests
             public override void OnSendingRequest(HttpMessage message)
             {
                 message.Request.Headers.Add(_headerName, _headerVaue);
+            }
+        }
+
+        private class TestOptions : ClientOptions
+        {
+        }
+
+        private class CustomResponseClassifier : ResponseClassifier
+        {
+            public override bool IsRetriableResponse(HttpMessage message)
+            {
+                return message.Response.Status == 500;
+            }
+
+            public override bool IsRetriableException(Exception exception)
+            {
+                return false;
+            }
+
+            public override bool IsErrorResponse(HttpMessage message)
+            {
+                return IsRetriableResponse(message);
             }
         }
         #endregion