+++ sdk/storage/Azure.Storage.Blobs/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added support for service vesrion 2021-02-12
 - Added support for listing system containers with BlobServiceClient.GetBlobContainers() and .GetBlobContainersAsync().
 - Added support for blob names container invalid XML characters.
+- Added support for updating the key encryption key on a client-side encrypted blob.
 - Fixed a bug where BlobClient.Upload() and UploadAsync() when using client-side encryption would modify the Dictionary instance passed by the caller for blob metadata.
 
 ## 12.11.0-beta.1 (2021-11-03)