+++ sdk/storage/Azure.Storage.Blobs/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release History
 
 ## 12.11.0-beta.2 (Unreleased)
-- Added support for service vesrion 2021-02-12
+- Added support for service version 2021-02-12
 - Added support for listing system containers with BlobServiceClient.GetBlobContainers() and .GetBlobContainersAsync().
 - Added support for blob names container invalid XML characters.
 - Added support for updating the key encryption key on a client-side encrypted blob.